requirements.txt
# General
Flask==2.2.5
Werkzeug==2.2.2
requests==2.31.0
pandas==1.3.3
numpy==1.21.2
openpyxl==3.0.9
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
mysqlclient==2.1.1
python-dotenv==1.0.0
APScheduler==3.10.3
Flask-Login==0.6.3
bcrypt==3.2.0

# Email
Flask-Mail==0.9.1

# Print
reportlab==3.6.1
PyPDF2==3.0.1
svglib==1.0.1
pymupdf==1.21.1

# MQTT
paho-mqtt

# Translation
Flask-Babel